sw/web/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.6
gunicorn>=21.0.0
django_redis
django-htmx
crispy-bulma
django-crispy-forms
markdown
